# runtime: use vDSO clock_gettime on linux/mips64x